feat(pre): complete rewrite for better testability (here: read)
