+++ bin/rlsr.js
@@ -1,4 +1,4 @@
 #! /usr/bin/env node
-const rlsr = require('rlsr');
+const rlsr = require('..');
 
 rlsr(process.argv[2]);
