+++ package.json
@@ -13,7 +13,8 @@
         "c": "git-cz",
         "test": "npm run lint && npm run unittest",
         "unittest": "jest",
-        "lint": "eslint ."
+        "lint": "eslint .",
+        "dryrun": "node ./bin/rlsr pre"
     },
     "repository": {
         "type": "git",
@@ -30,17 +31,17 @@
     "devDependencies": {
         "commitizen": "2.9.6",
         "cz-conventional-changelog": "2.0.0",
-        "jest": "^19.0.2"
+        "jest": "^20.0.4"
     },
     "dependencies": {
         "conventional-commits-parser": "1.3.0",
-        "eslint": "^3.18.0",
+        "eslint": "^4.0.0",
         "git-latest-semver-tag": "1.0.2",
-        "git-raw-commits": "1.1.2",
-        "glob": "7.1.1",
-        "npmlog": "4.0.2",
-        "ramda": "0.23.0",
-        "semistandard": "^10.0.0",
+        "git-raw-commits": "^1.2.0",
+        "glob": "^7.1.2",
+        "npmlog": "^4.1.0",
+        "ramda": "^0.24.1",
+        "semistandard": "^11.0.0",
         "semver": "5.3.0"
     },
     "config": {
