+++ src/transform/index.js
@@ -1,0 +1,1 @@
+module.exports = env => Promise.resolve(env);
