+++ src/write/index.js
@@ -1,0 +1,1 @@
+module.exports = env => console.log(env);
